backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==4.2.19
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
idna==3.10
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
django-extensions>=4.0
